Moves CollapsableList to separate file
